Add dumping macros and keybindings to morgue files.
This commit adds the possibility of dumping macros and custom
keybindings in the morgue file, useful for making sure the community
rules for realtime speedrunning have been followed.
It basically uses the same code which is used for writing the
macro.txt file, so the output isn't very readable, but it's good enough
to check what macros and keybindings have been used in a game.
It does not track changed or cleared macros, it only dumps what is
currently active at the time of death/win. A player could therefore
cheat by clearing their macros at some point.
It works even if the player says 'no' to the 'Save macros?" prompt.